migrate refactored classes to pytorch/sparsification
