+++ tests/sparseml/pytorch/sparsification/pruning/test_modifier_pruning_constant.py
@@ -16,13 +16,15 @@ import os
 
 import pytest
 
-from sparseml.pytorch.optim.pruning import ConstantPruningModifier
+from sparseml.pytorch.sparsification.pruning import ConstantPruningModifier
 from tests.sparseml.pytorch.helpers import LinearNet
-from tests.sparseml.pytorch.optim.pruning.helpers import state_dict_save_load_test
 from tests.sparseml.pytorch.optim.test_modifier import (
     ScheduledModifierTest,
     create_optim_adam,
     create_optim_sgd,
+)
+from tests.sparseml.pytorch.sparsification.pruning.helpers import (
+    state_dict_save_load_test,
 )
